refactor: move MakeHTTPClient to pkg adminapi
