+++ pkg/adminapi/header_roundtripper.go
@@ -1,4 +1,4 @@
-package main
+package adminapi
 
 import (
 	"net/http"
